timbuctoo-core/src/main/java/nl/knaw/huygens/timbuctoo/validation/RelationDuplicationValidator.java
package nl.knaw.huygens.timbuctoo.validation;

/*
 * #%L
 * Timbuctoo core
 * =======
 * Copyright (C) 2012 - 2014 Huygens ING
 * =======
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as
 * published by the Free Software Foundation, either version 3 of the 
 * License, or (at your option) any later version.
 * 
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 * 
 * You should have received a copy of the GNU General Public 
 * License along with this program.  If not, see
 * <http://www.gnu.org/licenses/gpl-3.0.html>.
 * #L%
 */

import nl.knaw.huygens.timbuctoo.model.Relation;
import nl.knaw.huygens.timbuctoo.storage.DuplicateException;
import nl.knaw.huygens.timbuctoo.storage.Repository;
import nl.knaw.huygens.timbuctoo.storage.StorageException;
import nl.knaw.huygens.timbuctoo.storage.ValidationException;

class RelationDuplicationValidator implements Validator<Relation> {

  private final Repository repository;

  public RelationDuplicationValidator(Repository repository) {
    this.repository = repository;
  }

  @Override
  public void validate(Relation relation) throws ValidationException {
    try {
      Relation entity = repository.findRelation(Relation.class, relation);
      if (entity != null) {
        throw new DuplicateException(entity.getId());
      }
    } catch (StorageException e) {
      throw new ValidationException("Error while validating", e);
    }
  }

}
